Settings Split into Retail & BC
